+++ src/index.js
@@ -1,5 +1,11 @@
 'use strict'
 
-exports.enums = require('./enums')
 exports.Client = require('./Client')
+exports.ClientAbstract = require('./ClientAbstract')
+exports.ClientPool = require('./ClientPool')
+exports.MessageQueue = require('./MessageQueue')
+exports.Room = require('./Room')
+exports.RoomManager = require('./RoomManager')
 exports.Server = require('./Server')
+exports.ServerClient = require('./ServerClient')
+exports.enums = require('./enums')